Rename PasswordEncoder to PasswordHasher and method encode to hash
